azalea::auto_tool
Function best_tool_in_hotbar_for_block

pub fn best_tool_in_hotbar_for_block(
    block: BlockState,
    menu: &Menu,
) -> BestToolResult

Returns the best tool in the hotbar for the given block.

Note that this doesn’t take into account whether the player is on the ground or in water, use accurate_best_tool_in_hotbar_for_block instead if you care about those things.