azalea
Module chat

Implementations of chat-related features.

Structs

ChatPlugin

ChatReceivedEvent

A client received a chat message packet.

SendChatEvent

Send a chat message (or command, if it starts with a slash) to the server.

SendChatKindEvent

Send a chat packet to the server of a specific kind (chat message or command). Usually you just want SendChatEvent instead.

Enums

ChatPacket

A chat packet, either a system message or a chat message.

ChatPacketKind

A kind of chat packet, either a chat message or a command.

Functions

handle_send_chat_event

handle_send_chat_kind_event