azalea
Module chunks

Used for Minecraft’s chunk batching introduced in 23w31a (1.20.2). It’s used for making the server spread out how often it sends us chunk packets depending on our receiving speed.

Structs

ChunkBatchFinishedEvent

ChunkBatchInfo

ChunkBatchStartEvent

ChunkPlugin

ChunkReceiveSpeedAccumulator

ReceiveChunkEvent

Functions

handle_chunk_batch_finished_event

handle_chunk_batch_start_event

handle_receive_chunk_events