azalea
Module events

Defines the Event enum and makes those events trigger when they’re sent in the ECS.

Structs

EventPlugin

LocalPlayerEvents

A component that contains an event sender for events that are only received by local players. The receiver for this is returned by Client::start_client.

Enums

Event

Something that happened in-game, such as a tick passing or chat message being sent.

Functions

add_player_listener

chat_listener

death_listener

disconnect_listener

init_listener

keepalive_listener

login_listener

packet_listener

remove_player_listener

tick_listener

update_player_listener