azalea
Module inventory

Modules

components

Representations of various inventory data structures in Minecraft.

item

operations

Structs

ClientSideCloseContainerEvent

Close a container without notifying the server.

CloseContainerEvent

Tell the server that we want to close a container.

ContainerClickEvent

DataComponentPatch

Inventory

A component present on all local players that have an inventory.

InventoryPlugin

InventorySet

ItemSlotData

An item in an inventory, with a count and NBT. Usually you want ItemSlot or azalea_registry::Item instead.

MenuOpenedEvent

Sent from the server when a menu (like a chest or crafting table) was opened by the client.

Player

SetContainerContentEvent

Sent from the server when the contents of a container are replaced. Usually triggered by the ContainerSetContent packet.

SetSelectedHotbarSlotEvent

SlotList

A fixed-size list of ItemSlots.

Enums

AnvilMenuLocation

BeaconMenuLocation

BlastFurnaceMenuLocation

BrewingStandMenuLocation

CartographyTableMenuLocation

Crafter3x3MenuLocation

CraftingMenuLocation

EnchantmentMenuLocation

FurnaceMenuLocation

Generic3x3MenuLocation

Generic9x1MenuLocation

Generic9x2MenuLocation

Generic9x3MenuLocation

Generic9x4MenuLocation

Generic9x5MenuLocation

Generic9x6MenuLocation

GrindstoneMenuLocation

HopperMenuLocation

ItemSlot

Either an item in an inventory or nothing.

LecternMenuLocation

LoomMenuLocation

Menu

A menu, which is a fixed collection of slots.

MenuLocation

MerchantMenuLocation

PlayerMenuLocation

ShulkerBoxMenuLocation

SmithingMenuLocation

SmokerMenuLocation

StonecutterMenuLocation

Functions

handle_client_side_close_container_event

handle_container_click_event