azalea::inventory
Module operations

Structs

CloneClick

Middle click, only defined for creative players in non-player inventories.

PickupAllClick

Double click

QuickCraftClick

SwapClick

Used when you press number keys or F in an inventory.

Enums

ClickOperation

ClickType

PickupClick

QuickCraftKind

QuickCraftStatus

QuickCraftStatusKind

QuickMoveClick

Shift click

ThrowClick