azalea
Module mining

Structs

AttackBlockEvent

A player left clicked on a block, used for stuff like interacting with note blocks.

FinishMiningBlockEvent

Sent when we completed mining a block.

LeftClickMine

A component that simulates the client holding down left click to mine the block that it’s facing, but this only interacts with blocks and not entities.

MineBlockPos

A component that stores the position of the block we’re currently mining.

MineBlockProgressEvent

MineBundle

A component bundle for players that can mine blocks.

MineDelay

A component that counts down until we start mining the next block.

MineItem

A component that contains the item we’re currently using to mine. If we’re not mining anything, it’ll be ItemSlot::Empty.

MinePlugin

A plugin that allows clients to break blocks in the world.

MineProgress

A component that stores the progress of the current mining operation. This is a value between 0 and 1.

MineTicks

A component that stores the number of ticks that we’ve been mining the same block for. This is a float even though it should only ever be a round number.

Mining

Information about the block we’re currently mining. This is only present if we’re currently mining a block.

MiningSet

StartMiningBlockEvent

Start mining the block at the given position.

StartMiningBlockWithDirectionEvent

StopMiningBlockEvent

Abort mining a block.

Functions

continue_mining_block

handle_finish_mining_block_event

handle_stop_mining_block_event