azalea::nearest_entity
Struct EntityFinder

pub struct EntityFinder<'w, 's, F = ()>
where
    F: QueryFilter + 'static,
{ /* private fields */ }

This system parameter can be used as a shorthand for quickly finding an entity, (or several) close to a given position.

This system parameter allows for additional filtering of entities based off of ECS marker components, such as With<>, Without<>, or Added<>, etc. All functions used by this system parameter instance will respect the applied filter.

use azalea::chat::SendChatEvent;
use azalea::nearest_entity::EntityFinder;
use azalea_entity::metadata::{Player, AbstractMonster};
use azalea_entity::LocalEntity;
use bevy_ecs::system::Query;
use bevy_ecs::prelude::{Entity, EventWriter};
use bevy_ecs::query::With;

/// All bots near aggressive mobs will scream in chat.
pub fn bots_near_aggressive_mobs(
    bots: Query<Entity, (With<LocalEntity>, With<Player>)>,
    entity_finder: EntityFinder<With<AbstractMonster>>,
    mut chat_events: EventWriter<SendChatEvent>,
) {
    for bot_id in bots.iter() {
        let Some(nearest) = entity_finder.nearest_to_entity(bot_id, 16.0) else {
            continue;
        };

        chat_events.send(SendChatEvent {
            entity: bot_id,
            content: String::from("Ahhh!"),
        });
    }
}

Implementations

impl<'a, F> EntityFinder<'_, '_, F>
where F: QueryFilter + 'static,

pub fn nearest_to_position( &'a self, position: &Position, instance_name: &InstanceName, max_distance: f64, ) -> Option<Entity>

Gets the nearest entity to the given position and world instance name. This method will return None if there are no entities within range. If multiple entities are within range, only the closest one is returned.

pub fn nearest_to_entity( &'a self, entity: Entity, max_distance: f64, ) -> Option<Entity>

Gets the nearest entity to the given entity. This method will return None if there are no entities within range. If multiple entities are within range, only the closest one is returned.

pub fn nearby_entities_to_position( &'a self, position: &'a Position, instance_name: &'a InstanceName, max_distance: f64, ) -> impl Iterator<Item = (Entity, f64)> + 'a

This function get an iterator over all nearby entities to the given position within the given maximum distance. The entities in this iterator are not returned in any specific order.

This function returns the Entity ID of nearby entities and their distance away.

pub fn nearby_entities_to_entity( &'a self, entity: Entity, max_distance: f64, ) -> impl Iterator<Item = (Entity, f64)> + 'a

This function get an iterator over all nearby entities to the given entity within the given maximum distance. The entities in this iterator are not returned in any specific order.

This function returns the Entity ID of nearby entities and their distance away.