azalea::packet_handling
Module login

Structs

IgnoreQueryIds

Plugins can add to this set if they want to handle a custom query packet themselves. This component removed after the login state ends.

LoginPacketEvent

An event that’s sent when we receive a login packet from the server. Note that if you want to handle this in a system, you must add .before(azalea::packet_handling::login::process_packet_events) to it because that system clears the events.

LoginSendPacketQueue

SendLoginPacketEvent

Event for sending a login packet to the server.

Functions

handle_send_packet_event

process_packet_events