azalea
Module pathfinder

A pathfinding plugin to make bots able to traverse the world.

Much of this code is based on Baritone.

Modules

astar

costs

goals

The goals that a pathfinder can try to reach.

mining

moves

simulation

Simulate the Minecraft world, currently only used for tests.

world

Structs

ComputePath

ExecutingPath

A component that’s present on clients that are actively following a pathfinder path.

GotoEvent

PathFoundEvent

Pathfinder

A component that makes this client able to pathfind.

PathfinderDebugParticles

A component that makes bots run /particle commands while pathfinding to show where they’re going. This requires the bots to have server operator permissions, and it’ll make them spam a lot of commands.

PathfinderPlugin

StopPathfindingEvent

Traits

PathfinderClientExt

Functions

add_default_pathfinder

call_successors_fn

check_for_path_obstruction

check_node_reached

check_path_obstructed

Checks whether the path has been obstructed, and returns Some(index) if it has been. The index is of the first obstructed node.

goto_listener

handle_stop_pathfinding_event

handle_tasks

path_found_listener

recalculate_if_has_goal_but_no_path

recalculate_near_end_of_path

stop_pathfinding_on_instance_change

tick_execute_path

timeout_movement