azalea::pathfinder
Module goals

The goals that a pathfinder can try to reach.

Structs

AndGoal

Try to reach both of the given goals.

AndGoals

Try to reach all the given goals.

BlockPosGoal

Move to the given block position.

InverseGoal

Do the opposite of the given goal.

OrGoal

Do either of the given goals, whichever is closer.

OrGoals

Do any of the given goals, whichever is closest.

RadiusGoal

Get within the given radius of the given position.

ReachBlockPosGoal

Move to a position where we can reach the given block.

XZGoal

Move to the given block position, ignoring the y-axis.

YGoal

Move to the given y coordinate.

Traits

Goal