azalea::pathfinder::moves
Struct ExecuteCtx

pub struct ExecuteCtx<'w1, 'w2, 'w3, 'w4, 'w5, 'w6, 'a> {

    pub entity: Entity,
    pub target: BlockPos,
    pub start: BlockPos,
    pub position: Vec3,
    pub physics: &'a Physics,
    pub is_currently_mining: bool,
    pub instance: Arc<RwLock<Instance>>,
    pub menu: Menu,
    pub look_at_events: &'a mut EventWriter<'w1, LookAtEvent>,
    pub sprint_events: &'a mut EventWriter<'w2, StartSprintEvent>,
    pub walk_events: &'a mut EventWriter<'w3, StartWalkEvent>,
    pub jump_events: &'a mut EventWriter<'w4, JumpEvent>,
    pub start_mining_events: &'a mut EventWriter<'w5, StartMiningBlockEvent>,
    pub set_selected_hotbar_slot_events: &'a mut EventWriter<'w6, SetSelectedHotbarSlotEvent>,

}

Fields
entity: Entitytarget: BlockPos

The node that we’re trying to reach.

start: BlockPos

The last node that we reached.

position: Vec3physics: &'a Physicsis_currently_mining: boolinstance: Arc<RwLock<Instance>>menu: Menulook_at_events: &'a mut EventWriter<'w1, LookAtEvent>sprint_events: &'a mut EventWriter<'w2, StartSprintEvent>walk_events: &'a mut EventWriter<'w3, StartWalkEvent>jump_events: &'a mut EventWriter<'w4, JumpEvent>start_mining_events: &'a mut EventWriter<'w5, StartMiningBlockEvent>set_selected_hotbar_slot_events: &'a mut EventWriter<'w6, SetSelectedHotbarSlotEvent>
Implementations

impl ExecuteCtx<'_, '_, '_, '_, '_, '_, '_>

pub fn look_at(&mut self, position: Vec3)

pub fn look_at_exact(&mut self, position: Vec3)

pub fn sprint(&mut self, direction: SprintDirection)

pub fn walk(&mut self, direction: WalkDirection)

pub fn jump(&mut self)

pub fn should_mine(&mut self, block: BlockPos) -> bool

Returns whether this block could be mined.

pub fn mine(&mut self, block: BlockPos) -> bool

Mine the block at the given position. Returns whether the block is being mined.

pub fn mine_while_at_start(&mut self, block: BlockPos) -> bool

Mine the given block, but make sure the player is standing at the start of the current node first.