azalea
Module prelude

The Azalea prelude. Things that are necessary for a bare-bones bot are re-exported here.

Re-exports

pub use crate::container::ContainerClientExt;

pub use crate::pathfinder::PathfinderClientExt;

pub use crate::ClientBuilder;

pub use crate::ecs as bevy_ecs;

Structs

Account

Something that can join Minecraft servers.

Client

Client has the things that a user interacting with the library will want.

GameTick

A Bevy schedule that runs every Minecraft game tick, i.e. every 50ms.

Enums

Event

Something that happened in-game, such as a tick passing or chat message being sent.

Traits

BotClientExt

Component

A data type that can be used to store data for an entity.

Resource

A type that can be inserted into a [World] as a singleton.

Derive Macros

Component

Resource