azalea::prelude
Trait BotClientExt

pub trait BotClientExt {
    // Required methods
    fn jump(&mut self);
    fn look_at(&mut self, pos: Vec3);
    fn get_tick_broadcaster(&self) -> Receiver<()>;
    fn mine(&mut self, position: BlockPos) -> impl Future<Output = ()> + Send;
}

Required Methods

fn jump(&mut self)

Queue a jump for the next tick.

fn look_at(&mut self, pos: Vec3)

Turn the bot’s head to look at the coordinate in the world.

fn get_tick_broadcaster(&self) -> Receiver<()>

Get a receiver that will receive a message every tick.

fn mine(&mut self, position: BlockPos) -> impl Future<Output = ()> + Send

Mine a block. This won’t turn the bot’s head towards the block, so if that’s necessary you’ll have to do that yourself with look_at.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.