azalea
Struct BlockPos

pub struct BlockPos {
    pub x: i32,
    pub y: i32,
    pub z: i32,
}

The coordinates of a block in the world. For entities (if the coordinate with decimals), use Vec3 instead.

Fields
x: i32y: i32z: i32
Implementations

impl BlockPos

pub fn new(x: i32, y: i32, z: i32) -> BlockPos

pub fn length_sqr(&self) -> i32

Get the distance of this vector to the origin by doing x^2 + y^2 + z^2.

pub fn distance_to_sqr(&self, other: &BlockPos) -> i32

Get the squared distance from this position to another position. Equivalent to (self - other).length_sqr().

pub fn horizontal_distance_sqr(&self) -> i32

pub fn horizontal_distance_to_sqr(&self, other: &BlockPos) -> i32

pub fn down(&self, y: i32) -> BlockPos

Return a new instance of this position with the y coordinate decreased by the given number.

pub fn up(&self, y: i32) -> BlockPos

Return a new instance of this position with the y coordinate increased by the given number.

pub fn north(&self, z: i32) -> BlockPos

Return a new instance of this position with the z coordinate subtracted by the given number.

pub fn east(&self, x: i32) -> BlockPos

Return a new instance of this position with the x coordinate increased by the given number.

pub fn south(&self, z: i32) -> BlockPos

Return a new instance of this position with the z coordinate increased by the given number.

pub fn west(&self, x: i32) -> BlockPos

Return a new instance of this position with the x coordinate subtracted by the given number.

pub fn dot(&self, other: BlockPos) -> i32

impl BlockPos

pub fn center(&self) -> Vec3

Get the absolute center of a block position by adding 0.5 to each coordinate.

pub fn to_vec3_floored(&self) -> Vec3

Convert the block position into a Vec3 without centering it.

pub fn length_manhattan(&self) -> u32

Get the distance of this vector from the origin by doing x + y + z.