azalea
Module swarm

Swarms are a way to conveniently control many bots.

Modules

prelude

A prelude that re-exports common swarm types in Azalea.

Structs

DefaultSwarmPlugins

This plugin group will add all the default plugins necessary for swarms to work.

NoSwarmState

A marker that can be used in place of a SwarmState in SwarmBuilder. You probably don’t need to use this manually since the compiler will infer it for you.

Swarm

A swarm is a way to conveniently control many bots at once, while also being able to control bots at an individual level when desired.

SwarmBuilder

Create a new Swarm.

Enums

SwarmEvent

An event about something that doesn’t have to do with a single bot.

Type Aliases

BoxSwarmHandleFn

SwarmHandleFn